tdf#81894 sc spelling: eliminate separate(inaccurate) check
ExecuteSpellPopup is quite accurate at identifying whether
the spelling at PosPixel is a spelling error or not.
If it is not a misspelled word, then it just silently returns.

If we make this a bDone flag, then it will just move
on to the normal right-click dialog.

There are a few additional scenarios that need to be
handled (mainly in terms of showing the edit-mode menu
instead of the full format menu, and also not
always returning the correct row/col when in edit mode).
I will do them in separate patches so that
each edge case can be clearly seen and handled.
They all rely on this patch being applied first.

This is the base patch. Too many things simply don't
work with the interim check, so I removed it.
(Part of the problem is that EditView kicks on and off,
and then a new CompleteOnlineSpelling needs to be run before
IsWrongSpelledWordAtPos can return a meaningful answer.)

This patch specifically fixes the problem of not
opening SpellPopup on a misspelled word in a wrapped-cell.
Steps to reproduce:
1.) Open LOv4304.ods from bug 81894 comment 6.
2.) Right-click on "spellng" in cell B12.

Before the patch, a full format popup was displayed
(not in edit mode) because IsSpellErrorAtPos returned
false, when it should have been true.

With this patch, spelling suggestions are provided.

As an example of an edge case, also notice:
3.) double-click on cell B4 (to enter edit mode)
4.) right click on "spellng" which appears to be in D4

Before the patch, nothing happened at all,
but with the patch at least the fall-back menu is displayed.
(EditView never switches out of cell B,
so SpellPopup is looking at wrong information.)

Change-Id: I91d413debc58de832dc7413034618e9e2bae3dd5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130297
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Eike Rathke <erack@redhat.com>