checked signedness of Sfx*IntItem types (part2)
See comments here
https://lists.freedesktop.org/archives/libreoffice/2022-January/088360.html

Make SfxItemSet::GetItemState check the type of the ppItem pointer
using TypedWhichId the same way we do for other calls.

Take the opportunity to name the new method GetItemIfSet and change
the signature to reduce the boilerplate at call sites..

This is just the initial commit, the bulk of the patch (the actual usage)
will be split into lots of smaller patches to aid bisecting for
bugs.

Change-Id: I6704059331787c8776f786af764b4ba3478a307b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128908
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>