tdf#80043 - Put selected text into hyperlink item
If the user selects an entire cell by a single click, and inserts an URL
via the hyperlink dialog, the entire content of the corresponding cell
will be replaced instead of inserting a duplicate, or appending the URL.
If the cell contains more than one URL or an URL including some
additional text within the cell, the hyperlink dialog will only
preselect the text field when opening. If this is not the case, it fill the URL field as well.

Change-Id: I9a10c3f302a6cf91a4d9dfadfbe775075eb00fad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131156
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>