tdf#147861 ww8export: DI_CUSTOM write fixed field, not plain txt
It looks like we just punt on fields, and try to avoid
round-tripping them as much as possible. And for good reason
since they are tough and poorly handled.

But since import of DI_CUSTOM has improved for DOC,
it would be nice to be able to round-trip it as well,
so that is what this allows.
It also opens the door to round-trip more field
situations as well.

This patch depends on the earlier related patch
that removes " (fixed)" from the FieldName.

RTF and DOCX also export OK, but import needs some love.

Change-Id: I32d9aa1fea12763cc30a79bcfa6d1d1e19d2c619
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131238
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>