SwNavigator: make resolved comments visible on Edit... if needed
Current behavior of comments members context menu item Edit..., when
both unresolved and resolved comments visibility options are off, is all
unresolved comments show before the cursor is placed in the comment
annotation window of the entry in the comments member list which
Edit... was selected. If the comment selected to edit is resolved the
cursor ends up in limbo. This fix makes resolved comments visible on
edit if needed.

Change-Id: Idf016493c70a150273fe558e4967284472684db3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131292
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>