ScriptForge - (SF_Root) differ loading of the localized interface
So far, the loading of the user interface in the user's
own language was done at the first invocation of
ScriptForge, usually via
- a CreateScriptService() method
- a direct call to any method, f.i. SF_Array.Append()

Now the same loading of the user interface in memory
is done at the first request of a localized text or message.

Due to the fact that since last commit
   https://gerrit.libreoffice.org/c/core/+/131255
only error messages contain localized texts,
the benefit is that correct scripts will not load any
unnecessary texts anymore.

The "Interface" variable has be renamed "LocalizedInterface"
for clarity in the SF_Root class module

No impact on documentation or Python scripts.

Change-Id: Ie654625540c0c8f8442e9ba36d38ecaab00c0eb1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131313
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>