related tdf#147861: cleanup + a function provides the name
This is not quite a NonFunctionalChange because it avoids
adding " (fixed)" to the name. But that is what I want
anyway for bug 147861, so this is an all-around good simplification.

Plus, the writerfilter loop doesn't need to run if an
inside clause will never be true anyway.

Putting this cleanup separately helps highlight the
actual fix and not clutter the real patches.

Change-Id: I12f345cc92627398ad00fe32b64b98b54e5ed14b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131413
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>