ScriptForge - (UI) size and position of the active window
The UI service receives next 4 new properties:
  Height
  Width
  X
  Y

They all return a Long value representing the size or position
of the active window. The active window does not need to be a
document, it may f.i. be the Basic IDE.

Those properties are read-only.
To modify the size or the position of the window,
use the the Resize() method.

They are implemented for use from Basic and Python scripts.

Change-Id: I0021663e39612f411cefa5c7ec9ec594a4cb6f39
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131444
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>