WeldEditView: Just pass new Size object as param
Just pass `aOutputSize` as parameter instead
of assigning that to `aSize` again and then
passing that one as parameter in
`WeldEditView::SetDrawingArea` and
overrides in derived classes.

`aSize.setHeight(aSize.Height())` can just be dropped
as an overly complicated way of keeping the
height as it is.

In `SidebarTextControl::SetDrawingArea`, `aOutputSize`
was already passed as param, so the new value of `aSize`
was ignored anyway.

Change-Id: I23192d3c5c85e4371a48774b3b8f854beb751b82
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131741
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>