Advanced Diagram support: Enhanced Ungroup for Diagram
A Diagram has a 'filler' BG object as 1st object in the
Group to guarantee the Diagram's dimensions. It has
no FillStyle or LineStyle and is Move/Size-Protected.

When Un-Grouping the Diagram and thus removing it's
Diagram-Functionality that filler-BG Object is not
useful and needs to be removed. Do that in an UNDO-
supporting way to allow seamless UNDO/REDO actions.

Change-Id: I1c5052de59dafb6ba1b35a68123bdabd10bf7885
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131930
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>