tdf#148235 Restore toolbar item to switch XForm to design mode
    commit 45b6f096e7ae86d0692ecdfd5b7069622d8a6efa
    Date:   Sun Oct 28 17:02:34 2018 +0100

        sw toolbars sync context toolbars between different apps

had replaced the "Form Design" toolbar item to switch
a complete XForm to design mode with an entry that
switches the form controls to design mode instead.
(The former one is labelled "Design Mode On/Off",
the latter one just "Design Mode" in the English UI.)

This meant that the XForm could no longer be switched
to design mode from there, unless the toolbar was
manually customized to add the entry back.

This brings the previous toolbar item back.
It also removes the one to switch the form controls
to design mode again, since that one is already contained
in the "Form Controls" toolbar, where it fits better.
(And having two different toolbar items with almost
identical labels but different functionality would
probably be rather confusing.)

Change-Id: Ia4c98dfa6ad8372eba08a9f08920153133a7f88d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132207
Reviewed-by: Andreas Kainz <kainz.a@gmail.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins