Extract 'switch' block out of WriteCustomGeometry
This is a follow up to commit 2029b2f6dd0109c5892e5ac5640022b31fe42fd2.
That commit has increased the line count of WriteCustomGeometry to more
than 500. This patch splits it to a main part of about 230 lines and
a new method for the previous 'switch' block of about 300 lines. That
makes the loops in the main part better readable.

Change-Id: Ied4378f54e7c8dc7965a5b1db15baf0b35f63f59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132274
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>