Remove duplicate DevGuide example FirstConnection
While cleaning up "LibreOffice DevGuide: Chapter 1 - First Steps" in
the TDF Wiki, I understood that only 3 examples were compiled via the
Makefile, but there were 4 java files. It turned out that the
FirstUnoContact.java and FirstConnection.java are essentially the
same. I removed the second one, as it was not built, and was not
referenced elsewhere.

Change-Id: I7df5dd9dfc3aae15de4484b4a3dd6272066a8683
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132457
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Hossein <hossein@libreoffice.org>