sw lok: fix missing cache invalidation in SwRedlineTable::Insert()
The trouble is that the FillRects() call in
SwRedlineTable::LOKRedlineNotification() builds a text portion list, but
that's not yet correct, and later we don't build a text portion list as
we already have one.

Fix this similar to the frame size problem by invalidating the cache
after we got our rectangles.

Change-Id: Ida759be418bc3706810d9774e060d06143893bb6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132521
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins