CppunitTest_sw_layoutwriter: avoid some a11y-based layout testing
The contents of SwTextFrames are currently dumped two times in the sw
xml layout dump: the first is misusing a11y's SwPortionHandler interface
for testing purposes (older approach), the second is visiting the
portions inside the text frame directly (newer approach).

The advantage of the second is that it can show information inside
multi-portions, and also extending it won't affect a11y. This commit
updates some asserts to use the second -- the first can be removed once
no asserts depend on it.

Change-Id: I5112946a964f338fc873ba13300023de9418e837
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132687
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>