tdf#148380 sw ms: import select fldLock'd fields as FIXED
Many fields we are locking just because we can't trust that
what is seen in MSWord is what will be seen in LO.

Of the unlocked fields which normally should be updated
(SAVEDATE and PRINTDATE), if MS Word secretly has also
locked them (via Ctrl-F11), then we also need to lock them.

Currently, on export these will be exported as plain text,
which IMHO is perfect because them we never need to mess
with these nasty fields again.

WriterFilter import was already handling the complicated FldChar's
fldLock OK, but nothing happened with the one for fldSimple.
Proving once again that a monkey can program, I randomly made
copy-changes to model.xml until I got the result I wanted...
The unit test has one of each.

Change-Id: Ia197794f4ea7e105b67cb1805c5def5347d7690d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133087
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>